Assert results tests are now able to use seedable rng
