+++ src/evaluate/roll.rs
@@ -1,5 +1,31 @@
 use crate::evaluate::dice_roll::to_notations;
 
+use super::expression::RolledExpression;
+
+#[derive(Debug, Clone, Copy)]
+#[repr(usize)]
+pub enum ModifierFlags {
+    Min,
+    Max,
+    ExplodingStandard,
+    ExplodingPenetrating,
+    ExplodingCompounding,
+    ExplodingPenetratingCompounding,
+    ReRoll,
+    ReRollOnce,
+    Unique,
+    UniqueOnce,
+    Drop,
+    TargetSuccess,
+    TargetFailure,
+    CriticalSuccess,
+    CriticalFailure,
+}
+
+pub const MODIFER_NOTATION: [&str; 16] = [
+    "^", "v", "!", "!p", "!!", "!!p", "r", "ro", "u", "uo", "d", "d", "*", "_", "**", "__",
+];
+
 #[derive(Debug, Clone, Copy)]
 pub struct Roll {
     pub(crate) value: i32,
@@ -80,29 +106,47 @@ pub enum RollOutputKind {
     TargetFailure,
 }
 
-#[derive(Debug, Clone, Copy)]
-#[repr(usize)]
-pub enum ModifierFlags {
-    Min,
-    Max,
-    ExplodingStandard,
-    ExplodingPenetrating,
-    ExplodingCompounding,
-    ExplodingPenetratingCompounding,
-    ReRoll,
-    ReRollOnce,
-    Unique,
-    UniqueOnce,
-    Drop,
-    TargetSuccess,
-    TargetFailure,
-    CriticalSuccess,
-    CriticalFailure,
+#[derive(Debug, Clone)]
+pub struct GroupRollOutput {
+    pub(crate) modifier_flags: Vec<u32>,
+    pub(crate) expressions: Vec<RolledExpression>,
+    pub(crate) kind: RollOutputKind,
 }
 
-pub const MODIFER_NOTATION: [&str; 16] = [
-    "^", "v", "!", "!p", "!!", "!!p", "r", "ro", "u", "uo", "d", "d", "*", "_", "**", "__",
-];
+impl GroupRollOutput {
+    #[rustfmt::skip]
+    pub fn value(&self) -> f64 {
+        let iter = self.expressions.iter().enumerate().filter(|(i, _)| {
+            !Self::was_modifier_applied(self.modifier_flags[*i], ModifierFlags::Drop as u8)
+        });
+
+        match self.kind {
+            RollOutputKind::Sum => iter.map(|(_, expr)| expr.value()).sum(),
+            RollOutputKind::TargetSuccess => iter
+                .map(|(i, _)| {
+                    let success = Self::was_modifier_applied(self.modifier_flags[i], ModifierFlags::TargetSuccess as u8);
+                    if success { 1.0 } else { 0.0 }
+                })
+                .sum(),
+            RollOutputKind::TargetFailure => iter
+                .map(|(i, _)| {
+                    let success = Self::was_modifier_applied(self.modifier_flags[i], ModifierFlags::TargetSuccess as u8);
+                    let failure = Self::was_modifier_applied(self.modifier_flags[i], ModifierFlags::TargetFailure as u8);
+                    if success { 1.0 } else if failure { -1.0 } else { 0.0 }
+                })
+                .sum(),
+        }
+    }
+
+    // NOTE: This is not a method of GroupRollOutput because otherwise I would not
+    // be allowed to mutably borrow self while iterating over self.expressions
+    pub fn set_modifier_flag(flags: &mut u32, flag: u8) {
+        *flags |= 1 << flag;
+    }
+    pub fn was_modifier_applied(flags: u32, flag: u8) -> bool {
+        (flags & (1 << flag)) != 0
+    }
+}
 
 impl PartialEq for Roll {
     fn eq(&self, other: &Self) -> bool {
@@ -147,5 +191,27 @@ impl std::fmt::Display for RollOutput {
             }
         }
         write!(f, "{}", str)
+    }
+}
+
+impl std::fmt::Display for GroupRollOutput {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        let group = self
+            .expressions
+            .iter()
+            .enumerate()
+            .map(|(ei, expr)| {
+                let mut notations = String::new();
+                for (ni, notation) in MODIFER_NOTATION.iter().enumerate() {
+                    if GroupRollOutput::was_modifier_applied(self.modifier_flags[ei], ni as u8) {
+                        notations.push_str(notation);
+                    }
+                }
+                format!("{expr}{notations}")
+            })
+            .collect::<Vec<_>>()
+            .join(", ");
+
+        write!(f, "{group}")
     }
 }
