+++ src/parse/dice.rs
@@ -1,10 +1,6 @@
 use winnow::{
     ascii::{dec_int, dec_uint, multispace0},
     combinator::{alt, cut_err, delimited, opt, preceded, repeat, separated_pair},
-    error::{
-        StrContext::{Expected, Label},
-        StrContextValue::{CharLiteral, StringLiteral},
-    },
     PResult, Parser,
 };
